Nutritionbot update (#390)
* updated user stat service

* calling nutritional service asynchronously

* linting issue missing semi colon
